requirements.txt
aiogram==3.15.0
SQLAlchemy==2.0.36
aiosqlite==0.20.0
aiofiles==24.1.0
greenlet==3.1.1
google-generativeai==0.8.3
python-dotenv==1.0.1
Pillow==11.1.0

